Modified reads2ref code to fix performance issue. Also, broke code up to add reads2pileups function.
